fix more test
